+++ src/data/restore.ts
@@ -132,11 +132,12 @@ const restoreElement = (
         fontSize,
         fontFamily,
         text: element.text ?? "",
+        rawText: element.rawText ?? "",
         baseline: element.baseline,
         textAlign: element.textAlign || DEFAULT_TEXT_ALIGN,
         verticalAlign: element.verticalAlign || DEFAULT_VERTICAL_ALIGN,
         containerId: element.containerId ?? null,
-        originalText: element.originalText ?? "",
+        originalText: element.originalText || element.text,
       });
     case "freedraw": {
       return restoreElementWithProperties(element, {