+++ src/tests/helpers/api.ts
@@ -88,6 +88,7 @@ export class API {
     opacity?: ExcalidrawGenericElement["opacity"];
     // text props
     text?: T extends "text" ? ExcalidrawTextElement["text"] : never;
+    rawText?: T extends "text" ? ExcalidrawTextElement["rawText"] : never;
     fontSize?: T extends "text" ? ExcalidrawTextElement["fontSize"] : never;
     fontFamily?: T extends "text" ? ExcalidrawTextElement["fontFamily"] : never;
     textAlign?: T extends "text" ? ExcalidrawTextElement["textAlign"] : never;
@@ -139,6 +140,7 @@ export class API {
         element = newTextElement({
           ...base,
           text: rest.text || "test",
+          rawText: rest.rawText || "test",
           fontSize: rest.fontSize ?? appState.currentItemFontSize,
           fontFamily: rest.fontFamily ?? appState.currentItemFontFamily,
           textAlign: rest.textAlign ?? appState.currentItemTextAlign,